feat(core): handle slices impossible to save on figma (#8)
* feat(core): handle slices impossible to save on figma

added an action to generate the theme page (with some preset of slices) and a Box component (which
contains all the variants). Added also a Sync action to update the existing Box variants according
to the current styles on the slices frames (and add some new if any)

wip #7

* feat(core): improved copy on a utils test

* fix(core): refactoring of getVariantCombinations function

* fix(core): changed the name of Theme and added a check to prevent creating it if already exist

changed name of Theme to "#Morfeo theme", extracted it to a constant and added a check to prevent
creating it in case it already exist. Added some tests to cover both cases

wip #7