feat(core): updated syncTheme to detect deleted slice variants and remove related component instances (#21)
* feat(core): updated syncTheme to detect deleted slice variants and remove related variants

implmented logic to keep trace of existing box variants and used it to detect any deleted slice
variants. In this way we can easily remove the related box variants and keep everything syncronised

closes #12

* fix(core): added some checks on syncTheme

now before updating box variants we check if all the slice frames are there and if there's at least
one variant for each of them. If not, the plugin will not update/add/delete anything

closes #12

* test(core): added some tests to cover syncTheme utils

closes #12

* fix(core): apply suggestion from code review

closes #12